requirements.txt
Flask
Flask-Cors
python-dotenv
pydantic
google-generativeai
gunicorn